Reorganize package in test to match main source package
