+++ build.gradle.kts
@@ -1,34 +1,50 @@
+/*
+ * Copyright 2026 Open Health Stack Foundation
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *       http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 import com.diffplug.gradle.spotless.SpotlessExtension
 
 plugins {
-    // this is necessary to avoid the plugins to be loaded multiple times
-    // in each subproject's classloader
-    alias(libs.plugins.androidApplication) apply false
-    alias(libs.plugins.androidLibrary) apply false
-    alias(libs.plugins.androidKotlinMultiplatformLibrary) apply false
-    alias(libs.plugins.composeCompiler) apply false
-    alias(libs.plugins.composeHotReload) apply false
-    alias(libs.plugins.composeMultiplatform) apply false
-    alias(libs.plugins.kotlinMultiplatform) apply false
-    alias(libs.plugins.spotless)
+  // this is necessary to avoid the plugins to be loaded multiple times
+  // in each subproject's classloader
+  alias(libs.plugins.androidApplication) apply false
+  alias(libs.plugins.androidLibrary) apply false
+  alias(libs.plugins.androidKotlinMultiplatformLibrary) apply false
+  alias(libs.plugins.composeCompiler) apply false
+  alias(libs.plugins.composeHotReload) apply false
+  alias(libs.plugins.composeMultiplatform) apply false
+  alias(libs.plugins.kotlinMultiplatform) apply false
+  alias(libs.plugins.spotless)
 }
 
 allprojects {
-    apply(plugin = rootProject.libs.plugins.spotless.get().pluginId)
+  apply(plugin = rootProject.libs.plugins.spotless.get().pluginId)
 
-    configure<SpotlessExtension> {
-        val ktfmtVersion = rootProject.libs.versions.ktfmt.get()
-        val licenseHeaderFile = rootProject.file("license-header.txt")
+  configure<SpotlessExtension> {
+    val ktfmtVersion = rootProject.libs.versions.ktfmt.get()
+    val licenseHeaderFile = rootProject.file("license-header.txt")
+    ratchetFrom = "origin/main"
 
-        kotlin {
-            target("src/**/*.kt")
-            ktfmt(ktfmtVersion).googleStyle()
-            licenseHeaderFile(licenseHeaderFile)
-        }
-        kotlinGradle {
-            target("*.gradle.kts")
-            ktfmt(ktfmtVersion).googleStyle()
-            licenseHeaderFile(licenseHeaderFile, "(^(?![\\/ ]\\*).*$)")
-        }
+    kotlin {
+      target("src/**/*.kt")
+      ktfmt(ktfmtVersion).googleStyle()
+      licenseHeaderFile(licenseHeaderFile)
     }
+    kotlinGradle {
+      target("*.gradle.kts")
+      ktfmt(ktfmtVersion).googleStyle()
+      licenseHeaderFile(licenseHeaderFile, "(^(?![\\/ ]\\*).*$)")
+    }
+  }
 }
