+++ ohs-player-library/src/commonMain/kotlin/dev/ohs/player/library/util/JsonUtil.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package dev.ohs.player.library.domain.util
+package dev.ohs.player.library.util
 
 import kotlinx.serialization.json.Json
