--- ohs-player-library/src/commonTest/kotlin/dev/ohs/player/library/PatientState.kt
@@ -1,37 +1,0 @@
-/*
- * Copyright 2026 Open Health Stack Foundation
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *       http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package dev.ohs.player.library
-
-import kotlinx.serialization.Serializable
-
-/**
- * Test state class representing extracted Patient fields.
- *
- * All fields are nullable with defaults. Each [ViewDefinition] populates only the fields it
- * declares — unmapped fields remain null.
- */
-@Serializable
-data class PatientState(
-  val patientId: String? = null,
-  val familyName: String? = null,
-  val fullName: String? = null,
-  val gender: String? = null,
-  val birthDate: String? = null,
-  val phone: String? = null,
-  val email: String? = null,
-  val city: String? = null,
-  val country: String? = null,
-)
